requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
Django==3.2.13
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.3.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
